EHPR-108 | Fix DB connection issue
